Title: 낮은 지점들, Time: 133ms, Memory: 2MB, Status: Passed - CodeTree

+++ 241014/낮은 지점들/lowest-points.cpp
@@ -19,7 +19,7 @@ int main() {
             spot[x] = min(spot[x], y);
     }
 
-    int sum = 0;
+    long long sum = 0;
     for(auto& pair : spot) {
         sum += pair.second;
     }
